requirements.txt
alembic==1.9.4
click==8.1.3
colorama==0.4.6
Flask==2.2.3
Flask-Login==0.6.2
greenlet==2.0.2
importlib-metadata==6.0.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
python-dotenv==0.21.1
SQLAlchemy==2.0.4
typing_extensions==4.5.0
Werkzeug==2.2.3
zipp==3.14.0
